requirements.txt
discord.py==2.2.3
watchfiles==0.19.0
python-dotenv==1.0.0
